# api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Les Canetons API",
        "version": "1.0.0",
...
    },
    "servers": [
        {
            "url": "https://lescanetons.org/api/v1",
            "description": "Production"
        }
    ],
...
    "paths": {
...
        "/events/{event}": {
            "get": {
                "operationId": "event.show",
                "description": "Any logged-in member; no permission is needed. Returns the event with\n`myAttendance`, the caller's own answer or `null`.\n\nWorks for a past event as well as an upcoming one, unlike the default\nlist. An unknown id answers `404`.",
                "summary": "Read one event",
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`EventResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "event.update",
                "description": "Requires `events.manage`. Send only the fields that change: an omitted\nfield is left alone, and an explicit `null` clears an optional one such\nas `attire` or `notes`. Returns the updated event, including the\ncaller's own `myAttendance`.\n\n`endsAt` must still come after the start. A request that sends a new end\nand no new start is compared against the stored `startsAt`, and fails\nagainst `endsAt` with `must_be_after` if it falls before it.\n\nClearing `registrationClosesAt` switches public registration off again.",
                "summary": "Correct an event already on the planning",
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEventRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`EventResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EventResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "400": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/Problem403"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            },
            "delete": {
                "operationId": "event.destroy",
                "description": "Requires `events.manage`. Deletes the event and everything attached to\nit: every attendance answer given for it, and every public registration\nbooked for it.\n\nAnswers `{\"ok\": true, \"attendanceDeleted\": n, \"registrationsDeleted\": n}`\nwith the counts of what went with it, so a confirmation can say what is\nabout to be discarded rather than merely asking again.",
                "summary": "Take an event off the planning",
                "tags": [
                    "Events"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean"
                                        },
                                        "attendanceDeleted": {
                                            "type": "integer",
                                            "minimum": 0
                                        },
                                        "registrationsDeleted": {
                                            "type": "integer",
                                            "minimum": 0
                                        }
                                    },
                                    "required": [
                                        "ok",
                                        "attendanceDeleted",
                                        "registrationsDeleted"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/Problem403"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            }
        },
...
        "/events/{event}/attendance": {
            "put": {
                "operationId": "attendance.update",
                "description": "Any logged-in member who is in a register; no permission is needed.\nAnswering is idempotent: sending an answer again replaces the previous\none rather than adding a second, and the response is always `200` with\nthe answer as it now stands, including its `recordedAt`.\n\n`status` is `yes` or `no`; anything else fails validation against that\nfield with `invalid_value`, which lists what is accepted.\n\nTaking back a `yes` costs a reason: if the stored answer is `yes` and\nthe incoming one is `no`, `note` is required, and a blank one fails\nvalidation against `note` with `required`. A first answer of `no`, and\na change from `no` to `yes`, cost nothing. Recording on somebody else's\nbehalf is exempt from this rule.\n\nA member who is in no register is refused with `403 not_answerable`:\nnothing is being asked of them, and no permission would change that.\n\nAnswering for yourself clears any mark saying the direction entered the\nanswer.",
                "summary": "Record your own answer for an event",
                "tags": [
                    "Attendance"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecordOwnAttendanceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AttendanceResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AttendanceResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "400": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            },
            "delete": {
                "operationId": "attendance.destroy",
                "description": "Any logged-in member; no permission is needed. Removes the answer\nentirely and returns the event to unanswered, which is a state a second\n`PUT` cannot express. Answers `{\"ok\": true}`, and taking back an answer\nthat is not there is not an error.\n\nThe window closes five minutes after the answer was last recorded;\nafter that it answers `409 answer_already_settled`. `recordedAt` on the\nanswer is what tells a client whether the window is still open, so it\ncan offer the undo or not rather than finding out from a refusal.",
                "summary": "Take back your own answer",
                "tags": [
                    "Attendance"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "ok"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            },
            "get": {
                "operationId": "attendance.index",
                "description": "Requires `attendance.view_all`. Returns every member who is answerable\nfor the event, ordered by name, whether or not they have replied: an\nentry's `attendance` is `null` for somebody who has not, which is what\nthis list is read for.\n\nAnswerable means being in a register. A member who is in none, such as\nsomebody who only organises, never appears here.\n\nEach entry carries the member's name and register alongside their\nanswer, and an answer says whether the direction entered it rather than\nthe member giving it themselves.",
                "summary": "List who is coming to an event",
                "tags": [
                    "Attendance"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of `ChaseListEntryResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChaseListEntryResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/Problem403"
                    }
                }
            }
        },
        "/events/{event}/attendance/{member}": {
            "put": {
                "operationId": "memberAttendance.update",
                "description": "Requires `attendance.record_for_others`, for the member who telephones\nthe committee instead of answering themselves. Idempotent: sending an\nanswer again replaces the previous one, and the response is always\n`200` with the answer as it now stands. It is marked as recorded by the\ndirection, so a list does not imply the member replied.\n\n`status` is `yes` or `no`; anything else fails validation against that\nfield with `invalid_value`. `note` is free text for whatever the member\nsaid, and stays optional even when the answer takes back a `yes`. A\nmember answering for themselves owes a reason for that; inventing one\non somebody else's behalf would put words in their mouth.\n\nRefuses `409 cannot_record_for_self` when the member named is the\ncaller, whose own answer has its own endpoint, and `403 not_answerable`\nwhen that member is in no register.",
                "summary": "Record an answer on a member's behalf",
                "tags": [
                    "Attendance"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "member",
                        "in": "path",
                        "required": true,
                        "description": "The member ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecordMemberAttendanceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AttendanceResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AttendanceResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "400": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/Problem403"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            },
            "delete": {
                "operationId": "memberAttendance.destroy",
                "description": "Requires `attendance.record_for_others`. Removes that member's answer\nfor the event entirely, returning them to unanswered. Answers\n`{\"ok\": true}`, and taking back an answer that is not there is not an\nerror.\n\nUnlike a member undoing their own answer, this is not limited to five\nminutes: correcting a mis-aimed entry an hour later is the case it\nexists for.\n\nRefuses `409 cannot_record_for_self` when the member named is the\ncaller. Use `DELETE /api/v1/events/{event}/attendance` for your own\nanswer.",
                "summary": "Take back a member's answer",
                "tags": [
                    "Attendance"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "member",
                        "in": "path",
                        "required": true,
                        "description": "The member ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "ok"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/Problem403"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RecordOwnAttendanceRequest": {
                "type": "object",
                "description": "A member's own answer to an event.\n\n`status` is `yes` or `no`; there is no \"maybe\". The endpoint is idempotent,\nso sending the same answer twice is fine and answering again replaces the\nprevious answer.\n\nOne conditional rule: taking back a `yes` costs a reason. If the stored\nanswer is `yes` and this request sends `no`, `note` becomes required and an\nempty one fails validation against `note` with `required`. Saying yes late,\nor repeating a `no`, needs no note.\n\nRecording an answer on somebody else's behalf uses\n`PUT /api/v1/events/{event}/attendance/{member}`, which does not ask for the\nreason.",
                "properties": {
                    "status": {
                        "type": "string",
                        "description": "Whether the member is coming. One of `yes` or `no`.",
                        "enum": [
                            "yes",
                            "no"
                        ]
                    },
                    "note": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Free text. Required when this request changes the member's own stored `yes` into a `no`; optional otherwise.",
                        "maxLength": 255
                    }
                },
                "required": [
                    "status"
                ],
                "title": "RecordOwnAttendanceRequest"
            },
...
        },
        "responses": {
...
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}